backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
requests>=2.31.0
